feature-0.6.1 adding model E ("E1", "E2")

+++ views.py
@@ -20,9 +20,9 @@ def profile():
 @views_bp.route('/simulate', methods=['POST'])
 def simulate():
     base_model = str(request.form["baseModel"])
-    index_start = int(request.form["indexStart"])
-    index_stop = int(request.form["indexStop"])
-    increment = int(request.form["increment"])
+    index_start = float(request.form["indexStart"])
+    index_stop = float(request.form["indexStop"])
+    increment = float(request.form["increment"])
     q = ModelRequestObj(base_model=base_model,
                         index_start=index_start,
                         index_stop=index_stop,
